+++ src/main.rs
@@ -74,7 +74,7 @@ use crate::{
         resolve_aabb,
         resources::SelectedEntity,
     },
-    hotkeys::{SHORTCUT_FOCUS, SHORTCUT_GAZE},
+    hotkeys::{SHORTCUT_FOCUS, SHORTCUT_GAZE, SHORTCUT_MAP_SWAP},
     input::InputState,
     map::MapDataList,
     map_resources::MapResource,
@@ -410,12 +410,16 @@ pub async fn main() -> anyhow::Result<()> {
     resources.insert(InputState::default());
     resources.insert(MapDataList {
         current_map: 0,
+        previous_map: 0,
+        updated: false,
         maps: vec![],
     });
     resources.insert(ScopeOverrides::default());
     resources.insert(DebugShapes::default());
     resources.insert(EnabledShaderOverrides::default());
-    resources.insert(RenderSettings::default());
+    let mut render_settings = RenderSettings::default();
+    render_settings.draw_crosshair = config::with(|cfg| cfg.render_settings.draw_crosshair);
+    resources.insert(render_settings);
     resources.insert(ShadowMapsResource::create(dcs.clone()));
     resources.insert(CurrentCubemap(None, None));
     resources.insert(ActivityGroupFilter::default());
@@ -690,6 +694,15 @@ pub async fn main() -> anyhow::Result<()> {
                         if d.is_finite() {
                             camera.focus(pos, 10.0);
                         }
+                    } else if gui
+                        .egui
+                        .input_mut(|i| i.consume_shortcut(&SHORTCUT_MAP_SWAP))
+                    {
+                        let mut maps = resources.get_mut::<MapDataList>().unwrap();
+
+                        (maps.current_map, maps.previous_map) =
+                            (maps.previous_map, maps.current_map);
+                        maps.updated = maps.previous_map != maps.current_map;
                     }
                 }
                 last_frame = Instant::now();
@@ -1100,6 +1113,9 @@ pub async fn main() -> anyhow::Result<()> {
                     c.resources.map_resource_label_background = gdb.map_resource_label_background;
                     c.resources.resource_distance_limit = gdb.map_resource_distance_limit_enabled;
                     c.resources.filters = resource_filters;
+
+                    let render_settings = resources.get_mut::<RenderSettings>().unwrap();
+                    c.render_settings.draw_crosshair = render_settings.draw_crosshair;
                 });
                 config::persist();
             }